refactor(runtime): isolate effect projections
Problem:
The complete command-effect change pushed several historically large files beyond the repository's no-growth lint budgets and duplicated focused tests inside an oversized ledger test file.

Root cause:
New effect projection, checkpoint comparison, and tool-plan state lived beside unrelated legacy responsibilities.

Fix:
Extract those responsibilities into focused files, compact internal helpers, and keep legacy adapters as thin documented projections without changing behavior.

Verification:
go test ./internal/shellsafe ./internal/permission ./internal/evidence ./internal/checkpoint ./internal/agent ./internal/recovery
make lint
