+++ internal/evidence/evidence.go
@@ -1575,114 +1575,6 @@ func ToolCallPaths(args json.RawMessage) []string {
 	return out
 }
 
-// ToolEffects projects the shared shell effect model onto the boundaries used
-// by delivery, TaskPolicy, workspace serialization, and batch recovery.
-type ToolEffects struct {
-	StateMutation      bool
-	WorkspaceMutation  bool
-	ContentMutation    bool
-	RepositoryMutation bool
-	Known              bool
-	Reason             string
-}
-
-// ClassifyToolCall returns durable effects for one concrete tool invocation.
-// Bash is always classified from its arguments even when a caller claims a
-// read-only schema bit, preventing a stale/proxy hint from hiding a writer.
-func ClassifyToolCall(toolName string, args json.RawMessage, readOnly bool) ToolEffects {
-	name := strings.ToLower(strings.TrimSpace(toolName))
-	if name == "bash" || name == "shell" {
-		effects, _ := ClassifyBashToolCall(args)
-		return effects
-	}
-	if readOnly {
-		return ToolEffects{Known: true}
-	}
-	if IsNonMutationMetaTool(toolName) {
-		return ToolEffects{Known: true}
-	}
-	switch name {
-	case "ask", "todo_write", "complete_step", "bash_output", "wait":
-		return ToolEffects{Known: true}
-	default:
-		return ToolEffects{
-			StateMutation:     true,
-			WorkspaceMutation: true,
-			ContentMutation:   true,
-			Known:             true,
-			Reason:            "workspace content write by a writer-capable tool",
-		}
-	}
-}
-
-// ClassifyBashToolCall parses a Bash invocation once and returns both durable
-// tool effects and the stricter permission-reader projection.
-func ClassifyBashToolCall(args json.RawMessage) (ToolEffects, bool) {
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(args, &fields); err != nil {
-		return unknownToolEffects("invalid shell arguments"), false
-	}
-	command := stringField(fields, "command")
-	effect := shellsafe.ClassifyBash(command)
-	permissionReader := effect.IsPermissionReader()
-	if effect.Certainty == shellsafe.EffectKnown {
-		return ToolEffects{
-			StateMutation:      effect.AnyMutation(),
-			WorkspaceMutation:  effect.WorkspaceMutation(),
-			ContentMutation:    effect.ContentMutation(),
-			RepositoryMutation: effect.RepositoryMutation(),
-			Known:              true,
-			Reason:             commandEffectReason(effect),
-		}, permissionReader
-	}
-	// Conventional verification commands are host-recognized separately from
-	// permission readers. They may execute code or fill caches, but do not create
-	// delivery content debt unless an output flag says so.
-	if !bashMayMutate(command) {
-		return ToolEffects{Known: true}, false
-	}
-	return unknownToolEffects(effect.Reason), false
-}
-
-func commandEffectReason(effect shellsafe.CommandEffect) string {
-	domain := ""
-	switch {
-	case effect.Writes&shellsafe.WriteWorkspaceContent != 0:
-		domain = "workspace content write"
-	case effect.Writes&shellsafe.WriteRepositoryMetadata != 0:
-		domain = "repository metadata write"
-	case effect.Writes&shellsafe.WriteHostState != 0:
-		domain = "host state write"
-	case effect.Writes&shellsafe.WriteExternalState != 0:
-		domain = "external state write"
-	}
-	if domain == "" {
-		return effect.Reason
-	}
-	if effect.CommandFamily == "" {
-		return domain
-	}
-	return domain + " by " + effect.CommandFamily
-}
-
-func unknownToolEffects(reason string) ToolEffects {
-	if strings.TrimSpace(reason) == "" {
-		reason = "tool effects are not statically known"
-	}
-	return ToolEffects{
-		StateMutation:     true,
-		WorkspaceMutation: true,
-		ContentMutation:   true,
-		Reason:            reason,
-	}
-}
-
-// ToolCallMutates is the compatibility projection for callers that need to
-// know whether any durable state may change.
-func ToolCallMutates(toolName string, args json.RawMessage, readOnly bool) bool {
-	return ClassifyToolCall(toolName, args, readOnly).StateMutation
-}
-
 // ToolCallRequiresDeliveryCriteria reports whether a call begins execution
 // work that needs an acceptance contract. Mutations always qualify; verification
 // commands also qualify even though they are intentionally not mutations.
